fix: FSRS Decimal coercion, course ownership check, bandit uses effective difficulty, task deduplication

+++ backend/app/models/scheduler.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import BigInteger, ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import ForeignKey, Integer, String, UniqueConstraint
 from sqlalchemy.dialects.postgresql import JSON, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -26,4 +26,4 @@ class SchedulerModel(UUIDPrimaryKeyMixin, TimestampMixin, Base):
     # "sm2" until SWITCHOVER_THRESHOLD reviews, then "fsrs"
     strategy: Mapped[str] = mapped_column(String(10), nullable=False, default="sm2")
     # Total review count across all cards for this user/course
-    review_count: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
+    review_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
